# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.2",
    "title": "WAFRN Backend",
    "description": "API routes for wafrn social network"
  },
  "servers": [
    {
      "url": "http://0.0.0.0:3000/"
    }
  ],
  "paths": {
...
    "/api/search/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/userSearch/{term}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "captchaResponse": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "nsfw": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "birthDate": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/editProfile": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/forgotPassword": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "captchaResponse": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/activateUser": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "code": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/resetPassword": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "example": "any"
                  },
                  "captchaResponse": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}